requirements.txt
asgiref==3.8.1
Django==4.2.19
django-model-utils==5.0.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
flake8==7.1.1
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
sqlparse==0.5.3
typing_extensions==4.12.2
